Refacto code
